requirements.txt
fastapi>=0.68.1
uvicorn>=0.15.0
pydantic>=1.8.2
numpy>=1.21.2
pandas>=1.3.3
scikit-learn>=1.0.0
tensorflow>=2.6.0
xgboost>=1.4.0
lightgbm>=3.2.1
python-dotenv>=0.19.0
joblib>=1.0.0